cffi: carry a cdata's address as an integer
`W_CData._ptr` is `rffi.CCHARP`, a raw-flavour pointer, so its kind is
integer.  Spelled `*mut u8` it landed in the reference bank instead, and the
`cdata.ptr as usize` at the `ctypefunc::call` boundary became a
`lltype.cast_ptr_to_int` the typer left as a residual call — a symbolic
funcbox, which aborts a sub-walk of `W_CData.call`.

The field, the `new_cdata*` constructors and `raw_alloc` now carry the
address as `usize`; the readers that dereference it cast at their call site.

The comment on `W_CData`'s immutable-field declaration is corrected to name
the reason `_ptr` stays undeclared: `do_exit` clears it, which `_do_exit`
does not.

+++ pyre/pyre-interpreter/src/module/_cffi_backend/cglob.rs
@@ -94,7 +94,7 @@ pub fn address(w_glob: PyObjectRef) -> Result<PyObjectRef, PyError> {
     let _ = roots.pin_root(w_glob);
     let ptr = fetch_global_var_addr(roots.get(glob_slot))?;
     let w_ctypeptr = newtype::new_pointer_type(glob_arg(roots.get(glob_slot))?.w_ctype)?;
-    Ok(cdataobj::new_cdata(ptr, w_ctypeptr))
+    Ok(cdataobj::new_cdata(ptr as usize, w_ctypeptr))
 }
 
 static GLOB_TYPE_OBJ: OnceLock<usize> = OnceLock::new();
